Changes to be committed: 	modified:   kvae/model/config.py 	modified:   kvae/model/model.py 	modified:   kvae/train/train_lightning.py
kvae/model/model.py:
Unification of VAE and Kalman Filter. Though still
needs testing, spetially the loss functions.

kvale/train/train_lightning.py:
Problems computing gradients. on_after_backward
calls clip_grad_norm_ to register the norm of the
gradients, that then is register in TensorBoard.
The norm is used to normalize the gradients if
they are too large. This prevents exploding
gradients. This type of behaviour is common
in RNNs and LSTMs. The value of the norm is
configurable in config.yaml as grad_clip_val.